+++ trell-android/app/src/main/java/com/example/trellandroid/data/repositories/MainRepository.kt
@@ -8,7 +8,7 @@ import com.example.trellandroid.utils.Result
 interface MainRepository {
     //actual functions which will call api
     suspend fun fetchAllVlogs() : Result<List<VlogResponse>>
-    suspend fun setInterestScore(vlogId: Double, score: Int = 1) : Result<Unit>
+    suspend fun setInterestScore(vlogId: Long, score: Int = 1) : Result<Unit>
 
     //demo functions that won't be calling any api in MVP but will be required in actual app
     suspend fun fetchAllComments(vlogId: Long) : Result<List<CommentResponse>>
